renaming rpcserver folder
