+++ lib/Xud.js
@@ -4,7 +4,7 @@ const DB = require('./db/DB');
 const OrderBook = require('./orderbook/OrderBook');
 const LndClient = require('./lndclient/LndClient');
 const RaidenClient = require('./raidenclient/RaidenClient');
-const RpcServer = require('./rpcserver/RpcServer');
+const RpcServer = require('./rpc/RpcServer');
 const P2P = require('./p2p/P2P');
 const P2PServer = require('./p2p/P2PServer');
