+++ lib/rpc/RpcMethods.js
@@ -54,9 +54,11 @@ class RpcMethods {
    */
   async placeOrder(params) {
     const { order } = params;
-    // temporary simple invoices until swaps are operational
-    const invoice = await this.lndClient.addInvoice(order.price * order.quantity);
-    order.invoice = invoice.payment_request;
+    if (!this.lndClient.isDisabled()) {
+      // temporary simple invoices until swaps are operational
+      const invoice = await this.lndClient.addInvoice(order.price * order.quantity);
+      order.invoice = invoice.payment_request;
+    }
     await this.orderbook.addOrder(order);
   }
